requirements.txt
asgiref==3.7.2
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.15.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2024.1
